requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
defusedxml==0.7.1
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.62.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.16.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.0
proto-plus==1.27.1
protobuf==5.29.6
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.54.0
tenacity==9.1.3
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
websockets==15.0.1
youtube-transcript-api==1.2.4
